Refined add alerts methods
