+++ src/components/Alerts/Alert/Alert.css
@@ -1,0 +1,3 @@
+.alert {
+    margin-bottom: 0px;
+}
