+++ src/store/actions/alerts.js
@@ -1,4 +1,4 @@
-import { ADD_ALERT, REMOVE_ALERT } from '../constants/index';
+import { ADD_ALERT } from '../constants/index';
 let i = 0;
 
 export function addAlert(options = {}) {
@@ -8,12 +8,5 @@ export function addAlert(options = {}) {
         id: i++
     },
     type: ADD_ALERT
-  };
-}
-
-export function removeAlert(id) {
-  return {
-    payload: id,
-    type: REMOVE_ALERT
   };
 }
