+++ src/store/constants/index.js
@@ -1,2 +1,1 @@
 export const ADD_ALERT = 'ADD_ALERT';
-export const REMOVE_ALERT = 'REMOVE_ALERT';
