+++ src/store/reducers/alerts.js
@@ -1,4 +1,4 @@
-import { ADD_ALERT, REMOVE_ALERT } from '../constants/index';
+import { ADD_ALERT } from '../constants/index';
 
 export default function alerts(state = [], action) {
     const { payload, type } = action;
@@ -6,9 +6,6 @@ export default function alerts(state = [], action) {
     switch (type) {
         case ADD_ALERT:
             return [payload, ...state];
-
-        case REMOVE_ALERT:
-            return state.filter(alert => alert.id !== payload);
 
         default:
             return state;
